sm_explain_policy_test.go
package opensearch

import (
	"testing"
)

func TestSmExplainPolicyBuildURL(t *testing.T) {

	client := setupTestClient(t)

	tests := []struct {
		Name         string
		ExpectedPath string
		ExpectErr    bool
	}{
		{
			"",
			"/_plugins/_sm/policies/_explain",
			false,
		},
		{
			"my-policy",
			"/_plugins/_sm/policies/my-policy/_explain",
			false,
		},
	}

	for i, test := range tests {
		builder := client.SmExplainPolicy(test.Name)
		err := builder.Validate()
		if err != nil {
			if !test.ExpectErr {
				t.Errorf("case #%d: %v", i+1, err)
				continue
			}
		} else {
			// err == nil
			if test.ExpectErr {
				t.Errorf("case #%d: expected error", i+1)
				continue
			}
			path, _, _ := builder.buildURL()
			if path != test.ExpectedPath {
				t.Errorf("case #%d: expected %q; got: %q", i+1, test.ExpectedPath, path)
			}
		}
	}
}
